requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.22.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0
Pillow>=9.0.0
medmnist>=2.2.0
